+++ src/StackExchange.Redis/ConnectionMultiplexer.cs
@@ -1370,10 +1370,11 @@ namespace StackExchange.Redis
 
         private void ActivateAllServers(ILogger? log)
         {
+            bool hasSubscriptions = GetSubscriptionsCount() != 0;
             foreach (var server in GetServerSnapshot())
             {
                 server.Activate(ConnectionType.Interactive, log);
-                if (server.SupportsSubscriptions && !server.KnowOrAssumeResp3())
+                if (hasSubscriptions && server.SupportsSubscriptions && !server.KnowOrAssumeResp3())
                 {
                     // Intentionally not logging the sub connection
                     server.Activate(ConnectionType.Subscription, null);